+++ ForgeMod/build.gradle
@@ -18,6 +18,7 @@ repositories {
 sourceSets {
     forge1122
     forge1710
+    launchWrapper
 }
 
 dependencies {
@@ -26,7 +27,10 @@ dependencies {
     implementation "org.apache.logging.log4j:log4j-api:2.8.1"
     //noinspection GradlePackageUpdate // suppress IDEA
     implementation "org.apache.logging.log4j:log4j-core:2.8.1"
-    implementation "net.minecraft:launchwrapper:1.12"
+
+    // launch wrapper dependencies
+    launchWrapperImplementation "net.minecraft:launchwrapper:1.12"
+    launchWrapperImplementation sourceSets.main.output
 
     // the dependency for forge 1.12.2 part
     forge1122Implementation "net.minecraftforge:forge:1.12.2-14.23.5.2856:universal"
@@ -45,7 +49,7 @@ jar {
 
     manifest {
         attributes([
-                "TweakClass": 'zone.rong.nukejndilookupfromlog4j.tweaker.NukeJndiLookupFromLog4j',
+                "TweakClass": 'zone.rong.nukejndilookupfromlog4j.tweaker.NukeJndiLookupFromLog4jTweaker',
         ])
     }
 }
